Update docker.md
I thought this could be a better video tutorial, because It has a person in the video explaining and showing how to use Docker. Maybe some people enjoy or learn more if they see an actual person. Also, thought the dog was really cute. Feel free to just decline my request though, because you already had a video selected.