Added online book for C++
Book is last updated in 2015 but main focus of book is understanding data structures (basics and more advanced). Also there are implementation for data structures with pointers and without pointers.  Book is made out of 3 parts. Basics, intermediate and advanced stuff so it's good for more advanced developers if they want to skip basics and take on intermediate or advanced part right away.